+++ src/parse_client_request/mod.rs
@@ -1,9 +1,8 @@
 //! Parse the client request and return a Subscription
 mod postgres;
 mod query;
-mod sse;
+
 mod subscription;
-mod ws;
 
 pub use self::postgres::PgPool;
 // TODO consider whether we can remove `Stream` from public API
@@ -11,3 +10,8 @@ pub use subscription::{Stream, Subscription, Timeline};
 
 //#[cfg(test)]
 pub use subscription::{Content, Reach};
+
+#[cfg(test)]
+mod sse_test;
+#[cfg(test)]
+mod ws_test;
